Rename nodebatcherprocessor to batchprocessor
This is for consistency reasons and to ensure configuration parameters
match directory structure.
